requirements.txt
streamlit
pandas
numpy
sentence-transformers
google-genai
python-dotenv
torch
